ui: First hack of a scrolling version

+++ src/include/rect.hh
@@ -4,8 +4,8 @@
 
 struct Rect
 {
-    int16_t x;
-    int16_t y;
-    uint16_t width;
-    uint16_t height;
+    uint32_t x;
+    uint32_t y;
+    uint32_t width;
+    uint32_t height;
 };
